# Switch back to a stable sort