@frankmcsherry 2018-07-11
There are places where the stable sort is def better, because it is a merge sort that looks for sorted runs and if they exist it is better than the pattern defeating quicksort (the unstable implementation). The stable sort has run-time guarantees, which is another mark it its favor. Seems good to me. :)