@lqd 2018-07-12
The hype update :) — numbers on unreleased software:
- the `LocationInsensitive` _leapfrog_ variant goes from 139ms to 127ms